requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
torch>=2.2.0
numpy==1.26.4
pandas>=2.2.0
scikit-learn>=1.4.0
tqdm>=4.66.0
python-docx>=1.1.0
PyPDF2>=3.0.1
